requirements.txt
fastapi
uvicorn
requests
xmltodict
python-dotenv
flask
